move upstream/downstream fixture
